requirements.txt
requests
pinecone-client
fiftyone
torch
torchvision
Pillow
streamlit